I made the requested refactors.
