+++ app/middlewares/passwordEncrypt.js
@@ -1,7 +1,7 @@
 const bcrypt = require('bcryptjs'),
   saltRounds = 10;
 
-module.exports = (request, response, next) => {
+module.exports = (request, response, next) =>
   bcrypt.hash(request.body.password, saltRounds, (err, hash) => {
     if (err) {
       response.status(503).send('We could not save your data, try again later');
@@ -10,4 +10,3 @@ module.exports = (request, response, next) => {
       next();
     }
   });
-};
