+++ app/routes.js
@@ -1,7 +1,7 @@
-const userController = require('./controllers/user'),
+const usersController = require('./controllers/users'),
   userValidator = require('./middlewares/userValidator'),
   passwordEncrypt = require('./middlewares/passwordEncrypt');
 
 exports.init = app => {
-  app.post('/users', userValidator, passwordEncrypt, userController.save);
+  app.post('/users', [userValidator, passwordEncrypt], usersController.save);
 };
